refactor: consolidate more integration tests (output-processor-memory, message-ordering, input-processors)
