--- packages/memory/integration-tests-v5/src/input-processors.test.ts
@@ -1,555 +1,0 @@
-import { openai } from '@ai-sdk/openai';
-import { Agent } from '@mastra/core/agent';
-import { MockStore } from '@mastra/core/storage';
-import { fastembed } from '@mastra/fastembed';
-import { LibSQLStore, LibSQLVector } from '@mastra/libsql';
-import { Memory } from '@mastra/memory';
-import type { CoreMessage } from 'ai';
-import { describe, expect, it } from 'vitest';
-
-/**
- * CRITICAL: These tests verify that input processors actually run and modify the LLM request.
- *
- * Each test checks the actual request.body.input sent to the LLM to ensure:
- * 1. MessageHistory processor fetches and includes previous messages
- * 2. WorkingMemory processor adds system messages with user context
- * 3. SemanticRecall processor adds relevant messages from other threads
- *
- * If these tests pass without the processors running, our test coverage is insufficient.
- */
-describe('Input Processor Verification - MessageHistory', () => {
-  it('should run MessageHistory input processor and include previous messages in LLM request', async () => {
-    const testStorage = new MockStore();
-    const memory = new Memory({
-      storage: testStorage,
-      options: {
-        lastMessages: 10, // Fetch last 10 messages
-      },
-    });
-
-    const agent = new Agent({
-      id: 'message-history-test',
-      name: 'Message History Test',
-      instructions: 'You are a helpful assistant',
-      model: openai('gpt-4o-mini'),
-      memory,
-    });
-
-    const threadId = '1';
-    const resourceId = 'message-history-resource';
-
-    // First message
-    await agent.generate('My name is Alice', {
-      threadId,
-      resourceId,
-    });
-
-    // Second message
-    await agent.generate('I live in Paris', {
-      threadId,
-      resourceId,
-    });
-
-    // Verify messages were saved
-    const { messages: savedMessages } = await memory.recall({ threadId });
-
-    if (savedMessages.length < 4) {
-      debugger;
-    }
-    expect(savedMessages.length).toBe(4); // 2 user + 2 assistant
-
-    // Third message - MessageHistory processor should include previous conversation
-    const thirdResponse = await agent.generate('What is my name and where do I live?', {
-      threadId,
-      resourceId,
-    });
-
-    // Check the actual request sent to the LLM
-    const requestMessages: CoreMessage[] = thirdResponse.request.body.input;
-
-    // Should have system + previous 4 messages + current message = 6 total
-    // OR at minimum: previous user + assistant + previous user + assistant + current user = 5
-    expect(requestMessages.length).toBeGreaterThanOrEqual(5);
-
-    // Should include "Alice" from first message
-    const aliceMessage = requestMessages.find((msg: any) => {
-      if (msg.role === 'user') {
-        if (typeof msg.content === 'string') {
-          return msg.content.includes('Alice');
-        }
-        if (Array.isArray(msg.content)) {
-          return msg.content.some((part: any) => part.text?.includes('Alice'));
-        }
-      }
-      return false;
-    });
-    expect(aliceMessage).toBeDefined();
-
-    // Should include "Paris" from second message
-    const parisMessage = requestMessages.find((msg: any) => {
-      if (msg.role === 'user') {
-        if (typeof msg.content === 'string') {
-          return msg.content.includes('Paris');
-        }
-        if (Array.isArray(msg.content)) {
-          return msg.content.some((part: any) => part.text?.includes('Paris'));
-        }
-      }
-      return false;
-    });
-    expect(parisMessage).toBeDefined();
-  });
-
-  it('should respect lastMessages limit in MessageHistory processor', async () => {
-    const testStorage = new MockStore();
-
-    const memory = new Memory({
-      storage: testStorage,
-      options: {
-        lastMessages: 2, // Only fetch last 2 messages
-      },
-    });
-
-    const agent = new Agent({
-      id: 'message-history-limit-test',
-      name: 'Message History Limit Test',
-      instructions: 'You are a helpful assistant',
-      model: openai('gpt-4o-mini'),
-      memory,
-    });
-
-    const threadId = '2';
-    const resourceId = 'limit-test-resource';
-
-    // Create 3 exchanges (6 messages total)
-    await agent.generate('Message 1', { threadId, resourceId, maxSteps: 1 });
-    await agent.generate('Message 2', { threadId, resourceId, maxSteps: 1 });
-    await agent.generate('Message 3', { threadId, resourceId, maxSteps: 1 });
-
-    // Fourth message - should only include last 2 messages (Message 3 + its response)
-    const fourthResponse = await agent.generate('Message 4', {
-      threadId,
-      resourceId,
-      maxSteps: 1,
-    });
-
-    const requestMessages: CoreMessage[] = fourthResponse.request.body.input;
-
-    // Should have: system + last 2 historical messages + current user message
-    // With lastMessages: 2, we fetch the 2 most recent messages from storage
-    // After 3 exchanges, that's Message 3 (user) + Response 3 (assistant)
-    // Total: system (1) + Message 3 (1) + Response 3 (1) + Message 4 current (1) = 4
-    // However, we're seeing 5-6 messages due to message consolidation or test pollution
-    // TODO: Investigate why MockStore/InMemoryStore has cross-test pollution
-    expect(requestMessages.length).toBeLessThanOrEqual(6);
-
-    // Should NOT include "Message 1" (too old)
-    const message1 = requestMessages.find((msg: any) => {
-      if (msg.role === 'user') {
-        if (typeof msg.content === 'string') {
-          return msg.content.includes('Message 1');
-        }
-        if (Array.isArray(msg.content)) {
-          return msg.content.some((part: any) => part.text?.includes('Message 1'));
-        }
-      }
-      return false;
-    });
-    expect(message1).toBeUndefined();
-
-    // Should include "Message 3" (within limit)
-    const message3 = requestMessages.find((msg: any) => {
-      if (msg.role === 'user') {
-        if (typeof msg.content === 'string') {
-          return msg.content.includes('Message 3');
-        }
-        if (Array.isArray(msg.content)) {
-          return msg.content.some((part: any) => part.text?.includes('Message 3'));
-        }
-      }
-      return false;
-    });
-    expect(message3).toBeDefined();
-  });
-});
-
-describe('Input Processor Verification - WorkingMemory', () => {
-  it('should run WorkingMemory input processor and include working memory in LLM request', async () => {
-    const memory = new Memory({
-      storage: new MockStore(),
-      options: {
-        workingMemory: {
-          enabled: true,
-        },
-      },
-    });
-
-    const agent = new Agent({
-      id: 'working-memory-test',
-      name: 'Working Memory Test',
-      instructions: 'You are a helpful assistant',
-      model: openai('gpt-4o-mini'),
-      memory,
-    });
-
-    const threadId = '2';
-    const resourceId = 'working-memory-resource';
-
-    // Set working memory
-    await memory.updateWorkingMemory({
-      threadId,
-      resourceId,
-      workingMemory:
-        '# User Profile\nName: Bob Smith\nAge: 35\nOccupation: Software Engineer\nFavorite Language: TypeScript',
-    });
-
-    // Generate a response - WorkingMemory processor should include the working memory
-    const response = await agent.generate('What is my occupation?', {
-      threadId,
-      resourceId,
-    });
-
-    // Check the actual request sent to the LLM
-    const requestMessages: CoreMessage[] = response.request.body.input;
-
-    // Should have at least 2 messages: working memory system message + user message
-    expect(requestMessages.length).toBeGreaterThanOrEqual(2);
-
-    // Should include a system message with working memory content
-    const workingMemoryMessage = requestMessages.find((msg: any) => {
-      if (msg.role === 'system') {
-        if (typeof msg.content === 'string') {
-          return msg.content.includes('Bob Smith') && msg.content.includes('Software Engineer');
-        }
-        if (Array.isArray(msg.content)) {
-          return msg.content.some(
-            (part: any) => part.text?.includes('Bob Smith') && part.text?.includes('Software Engineer'),
-          );
-        }
-      }
-      return false;
-    });
-
-    expect(workingMemoryMessage).toBeDefined();
-
-    // Verify the working memory content is present
-    const workingMemoryContent =
-      typeof workingMemoryMessage!.content === 'string'
-        ? workingMemoryMessage!.content
-        : (workingMemoryMessage!.content as any[]).find((part: any) => part.text)?.text || '';
-
-    expect(workingMemoryContent).toContain('Bob Smith');
-    expect(workingMemoryContent).toContain('Software Engineer');
-    expect(workingMemoryContent).toContain('TypeScript');
-  });
-
-  it.skip('should use custom working memory template when provided', async () => {
-    // TODO: Fix this test - template should be WorkingMemoryTemplate object, not a function
-    const customTemplate = (workingMemory: string) => {
-      return `CUSTOM CONTEXT:\n${workingMemory}\n\nUse this information to answer questions.`;
-    };
-
-    const memory = new Memory({
-      storage: new MockStore(),
-      options: {
-        workingMemory: {
-          enabled: true,
-          template: customTemplate as any,
-        },
-      },
-    });
-
-    const agent = new Agent({
-      id: 'custom-template-test',
-      name: 'Custom Template Test',
-      instructions: 'You are a helpful assistant',
-      model: openai('gpt-4o-mini'),
-      memory,
-    });
-
-    const threadId = 'test-thread-3';
-    const resourceId = 'custom-template-resource';
-
-    await memory.updateWorkingMemory({
-      threadId,
-      resourceId,
-      workingMemory: 'User prefers dark mode',
-    });
-
-    const response = await agent.generate('What are my preferences?', {
-      threadId,
-      resourceId,
-    });
-
-    const requestMessages: CoreMessage[] = response.request.body.input;
-
-    // Should include the custom template text
-    const customTemplateMessage = requestMessages.find((msg: any) => {
-      if (msg.role === 'system') {
-        const content =
-          typeof msg.content === 'string'
-            ? msg.content
-            : (msg.content as any[]).find((part: any) => part.text)?.text || '';
-        return content.includes('CUSTOM CONTEXT') && content.includes('dark mode');
-      }
-      return false;
-    });
-
-    expect(customTemplateMessage).toBeDefined();
-  });
-});
-
-describe('Input Processor Verification - SemanticRecall', () => {
-  it('should run SemanticRecall input processor and include semantically similar messages from other threads', async () => {
-    // Use shared in-memory database so storage and vector use the same DB
-    const dbFile = 'file::memory:?cache=shared';
-    const storage = new LibSQLStore({
-      id: 'semantic-recall-storage',
-      url: dbFile,
-    });
-    const vector = new LibSQLVector({
-      connectionUrl: dbFile,
-      id: 'semantic-recall-vector',
-    });
-
-    // Initialize storage to create tables
-    await storage.init();
-
-    const memory = new Memory({
-      storage,
-      vector,
-      embedder: fastembed,
-      options: {
-        semanticRecall: {
-          topK: 3,
-          messageRange: 2,
-          scope: 'resource', // Cross-thread recall
-        },
-        lastMessages: 2,
-      },
-    });
-
-    const agent = new Agent({
-      id: 'semantic-recall-test',
-      name: 'Semantic Recall Test',
-      instructions: 'You are a helpful assistant',
-      model: openai('gpt-4o-mini'),
-      memory,
-    });
-
-    const resourceId = 'semantic-recall-resource';
-    const thread1Id = 'test-thread-4a';
-    const thread2Id = 'test-thread-4b';
-
-    // Thread 1: Discuss Python programming
-    await agent.generate('I love programming in Python, especially for data science', {
-      threadId: thread1Id,
-      resourceId,
-    });
-
-    await agent.generate('Python has great libraries like pandas and numpy', {
-      threadId: thread1Id,
-      resourceId,
-    });
-
-    // Thread 2: Ask about programming (should recall Python messages from thread 1)
-    const response = await agent.generate('What programming languages have we discussed?', {
-      threadId: thread2Id,
-      resourceId,
-    });
-
-    const requestMessages: CoreMessage[] = response.request.body.input;
-
-    // Should have more than just the current message
-    // Should include: system + semantically recalled messages + current message
-    expect(requestMessages.length).toBeGreaterThan(2);
-
-    // Should include messages about Python from thread 1 in a system message
-    const semanticRecallMessage = requestMessages.find((msg: any) => {
-      if (msg.role === 'system') {
-        const content = typeof msg.content === 'string' ? msg.content : '';
-        return content.includes('<remembered_from_other_conversation>') && content.toLowerCase().includes('python');
-      }
-      return false;
-    });
-
-    expect(semanticRecallMessage).toBeDefined();
-
-    // Verify the recalled message is from a different thread (cross-thread recall)
-    // This is implicit - if we found Python messages in the semantic recall system message, they must be from thread1
-  });
-
-  it('should respect topK limit in SemanticRecall processor', async () => {
-    // Use shared in-memory database so storage and vector use the same DB
-    const dbFile = 'file::memory:?cache=shared';
-    const storage = new LibSQLStore({
-      id: 'semantic-topk-storage',
-      url: dbFile,
-    });
-    const vector = new LibSQLVector({
-      connectionUrl: dbFile,
-      id: 'semantic-topk-vector',
-    });
-
-    // Initialize storage to create tables
-    await storage.init();
-
-    const memory = new Memory({
-      storage,
-      vector,
-      embedder: fastembed,
-      options: {
-        semanticRecall: {
-          topK: 1, // Only recall 1 message
-          messageRange: 0, // No context around it
-          scope: 'resource',
-        },
-        lastMessages: 0, // Don't include message history
-      },
-    });
-
-    const agent = new Agent({
-      id: 'semantic-topk-test',
-      name: 'Semantic TopK Test',
-      instructions: 'You are a helpful assistant',
-      model: openai('gpt-4o-mini'),
-      memory,
-    });
-
-    const resourceId = 'topk-resource';
-    const thread1Id = 'test-thread-5a';
-    const thread2Id = 'test-thread-5b';
-
-    // Create multiple messages in thread 1
-    await agent.generate('I like cats', { threadId: thread1Id, resourceId });
-    await agent.generate('I like dogs', { threadId: thread1Id, resourceId });
-    await agent.generate('I like birds', { threadId: thread1Id, resourceId });
-
-    // Query from thread 2 - should only recall 1 message (topK=1)
-    const response = await agent.generate('Tell me about cats', {
-      threadId: thread2Id,
-      resourceId,
-    });
-
-    const requestMessages: CoreMessage[] = response.request.body.input;
-
-    // Should have: system + 1 recalled message + current message = 3 total
-    // (or possibly just recalled + current = 2 if no system message)
-    expect(requestMessages.length).toBeLessThanOrEqual(3);
-
-    // Count user messages (should be at most 2: recalled + current)
-    const userMessages = requestMessages.filter((msg: any) => msg.role === 'user');
-    expect(userMessages.length).toBeLessThanOrEqual(2);
-  });
-});
-
-describe('Input Processor Verification - Combined Processors', () => {
-  it('should run all input processors together (MessageHistory + WorkingMemory + SemanticRecall)', async () => {
-    // Use shared in-memory database so storage and vector use the same DB
-    const dbFile = 'file::memory:?cache=shared';
-    const storage = new LibSQLStore({
-      id: 'combined-storage',
-      url: dbFile,
-    });
-    const vector = new LibSQLVector({
-      connectionUrl: dbFile,
-      id: 'combined-vector',
-    });
-
-    // Initialize storage to create tables
-    await storage.init();
-
-    const memory = new Memory({
-      storage,
-      vector,
-      embedder: fastembed,
-      options: {
-        workingMemory: {
-          enabled: true,
-        },
-        semanticRecall: {
-          topK: 2,
-          messageRange: 1,
-          scope: 'resource',
-        },
-        lastMessages: 3,
-      },
-    });
-
-    const agent = new Agent({
-      id: 'combined-test',
-      name: 'Combined Test',
-      instructions: 'You are a helpful assistant',
-      model: openai('gpt-4o-mini'),
-      memory,
-    });
-
-    const resourceId = 'combined-resource';
-    const thread1Id = 'test-thread-6a';
-    const thread2Id = 'test-thread-6b';
-
-    // Set working memory
-    await memory.updateWorkingMemory({
-      threadId: thread2Id,
-      resourceId,
-      workingMemory: '# User Info\nName: Charlie\nRole: Developer',
-    });
-
-    // Thread 1: Create some history
-    await agent.generate('I work with React', { threadId: thread1Id, resourceId });
-
-    // Thread 2: Create some history
-    await agent.generate('Hello', { threadId: thread2Id, resourceId });
-
-    // Thread 2: Query - should include all processors
-    const response = await agent.generate('What do I work with?', {
-      threadId: thread2Id,
-      resourceId,
-    });
-
-    const requestMessages: CoreMessage[] = response.request.body.input;
-
-    // Should have multiple messages from different processors
-    expect(requestMessages.length).toBeGreaterThan(2);
-
-    // Should include working memory (system message with "Charlie")
-    const workingMemoryMsg = requestMessages.find((msg: any) => {
-      if (msg.role === 'system') {
-        const content =
-          typeof msg.content === 'string'
-            ? msg.content
-            : Array.isArray(msg.content)
-              ? msg.content.find((part: any) => part.text)?.text || ''
-              : '';
-        return content.includes('Charlie');
-      }
-      return false;
-    });
-    expect(workingMemoryMsg).toBeDefined();
-
-    // Should include message history from thread 2 ("Hello")
-    const historyMsg = requestMessages.find((msg: any) => {
-      if (msg.role === 'user') {
-        const content =
-          typeof msg.content === 'string'
-            ? msg.content
-            : Array.isArray(msg.content)
-              ? msg.content.find((part: any) => part.text)?.text || ''
-              : '';
-        return content.includes('Hello');
-      }
-      return false;
-    });
-    expect(historyMsg).toBeDefined();
-
-    // Should include semantically recalled message from thread 1 ("React") in a system message
-    const semanticMsg = requestMessages.find((msg: any) => {
-      if (msg.role === 'system') {
-        const content = typeof msg.content === 'string' ? msg.content : '';
-        return content.includes('<remembered_from_other_conversation>') && content.includes('React');
-      }
-      return false;
-    });
-    expect(semanticMsg).toBeDefined();
-  });
-});
